requirements.txt
fastapi
uvicorn
influxdb-client
python-dotenv
jinja2